Removes test since there is no longer control of async operations
